Refactor nav logic into seperate files and add authenticated check
