+++ App.tsx
@@ -1,26 +1,9 @@
 import { Provider } from "react-redux";
 import { store } from "./src/redux/store";
-import { View, Text, Button, LogBox } from "react-native";
-import { NavigationContainer } from "@react-navigation/native";
+import { View, Text, LogBox } from "react-native";
 import { loadFonts } from "./src/utils/loadFonts";
-import HomeScreen from "./src/screens/mainScreens/HomeScreen";
-import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
-import ComposersScreen from "./src/screens/mainScreens/ComposersScreen";
-import SheetsScreen from "./src/screens/mainScreens/SheetsScreen";
-import SettingsScreen from "./src/screens/mainScreens/SettingsScreen";
-import CollectionsScreen from "./src/screens/mainScreens/CollectionsScreen";
-import MaterialCommunityIcons from "react-native-vector-icons/MaterialCommunityIcons";
-import Ionicons from "react-native-vector-icons/Ionicons";
-import { colors, globalStyles } from "./src/constants/GlobalStyleSheet";
-import { TabScreenText } from "./src/components/utils/TabScreenText";
 import axios from "axios";
-import { createStackNavigator } from "@react-navigation/stack";
-import SheetScreen from "./src/screens/functionalScreens/SheetScreen";
-import { RootStackParamList } from "./src/screens/RootStackParams";
-import SignInScreen from "./src/screens/authScreens/SignInScreen";
-
-const Tab = createBottomTabNavigator<RootStackParamList>();
-const Stack = createStackNavigator<RootStackParamList>();
+import MainNav from "./src/navigator/MainNav";
 
 export default function App() {
   axios.defaults.baseURL = "http://192.168.0.65:8080/api";
@@ -40,133 +23,7 @@ export default function App() {
 
   return (
     <Provider store={store}>
-      <NavigationContainer>
-        <Stack.Navigator
-          initialRouteName="Home"
-          screenOptions={{
-            headerMode: "screen",
-            headerTintColor: "white",
-            headerStyle: { backgroundColor: colors.BLUE8 },
-          }}
-        >
-          <Stack.Screen
-            name="Home"
-            component={TabNavigator}
-            options={{ headerShown: false }}
-          />
-          <Stack.Screen
-            name="Sheet"
-            component={SheetScreen}
-            options={({ route }) => ({
-              title: `Sheet - ${route.params.sheetName}`,
-            })}
-          />
-          <Stack.Screen
-            name="SignIn"
-            component={SignInScreen}
-            options={{
-              headerShown: false,
-              gestureEnabled: false,
-            }}
-          />
-        </Stack.Navigator>
-      </NavigationContainer>
+      <MainNav />
     </Provider>
-  );
-}
-
-function TabNavigator() {
-  return (
-    <Tab.Navigator
-      screenOptions={{
-        headerShown: false,
-        tabBarStyle: {
-          borderTopWidth: 1,
-          backgroundColor: "white",
-        },
-      }}
-    >
-      <Tab.Screen
-        name="HomeFeed"
-        component={HomeScreen}
-        options={{
-          tabBarLabel: ({ focused }) => (
-            <TabScreenText focused={focused} label="Home" />
-          ),
-          tabBarIcon: ({ focused, size }) => (
-            <Ionicons
-              name="home"
-              color={focused ? colors.PRIMARY_LIGHT : colors.GRAY8}
-              size={size}
-            />
-          ),
-        }}
-      />
-      <Tab.Screen
-        name="Sheets"
-        component={SheetsScreen}
-        options={{
-          tabBarLabel: ({ focused }) => (
-            <TabScreenText focused={focused} label="Sheets" />
-          ),
-          tabBarIcon: ({ focused, size }) => (
-            <MaterialCommunityIcons
-              name="library"
-              color={focused ? colors.PRIMARY_LIGHT : colors.GRAY8}
-              size={size + 3}
-            />
-          ),
-        }}
-      />
-      <Tab.Screen
-        name="Composers"
-        component={ComposersScreen}
-        options={{
-          tabBarLabel: ({ focused }) => (
-            <TabScreenText focused={focused} label="Composers" />
-          ),
-          tabBarIcon: ({ focused, size }) => (
-            <Ionicons
-              name="people"
-              color={focused ? colors.PRIMARY_LIGHT : colors.GRAY8}
-              size={size + 3}
-            />
-          ),
-        }}
-      />
-      {/* Not implemented yet:
-      <Tab.Screen
-        name="Collections"
-        component={CollectionsScreen}
-        options={{
-          tabBarLabel: ({ focused }) => (
-            <TabScreenText focused={focused} label="Collections" />
-          ),
-          tabBarIcon: ({ focused, size }) => (
-            <Ionicons
-              name="md-bookmarks"
-              color={focused ? colors.PRIMARY_LIGHT : colors.GRAY8}
-              size={size - 3}
-            />
-          ),
-        }}
-      />*/}
-      <Tab.Screen
-        name="Settings"
-        component={SettingsScreen}
-        options={{
-          tabBarLabel: ({ focused }) => (
-            <TabScreenText focused={focused} label="Settings" />
-          ),
-          tabBarIcon: ({ focused, size }) => (
-            <Ionicons
-              name="cog"
-              color={focused ? colors.PRIMARY_LIGHT : colors.GRAY8}
-              size={size + 5}
-            />
-          ),
-        }}
-      />
-    </Tab.Navigator>
   );
 }
