--- src/redux/slicers/counter/counterAPI.ts
@@ -1,5 +1,0 @@
-export function fetchCount(currentCount = 0, amount = 1) {
-  return new Promise<{ data: number }>((resolve) =>
-    setTimeout(() => resolve({ data: currentCount + amount }), 1000)
-  );
-}
